+++ src/router/index.js
@@ -17,6 +17,12 @@ const router = createRouter({
       // which is lazy-loaded when the route is visited.
       component: () => import('@/views/AboutView.vue'),
     },
+    {
+      path: '/result',
+      name: 'result',
+      props: { default: true, sidebar: false },
+      component: () => import('@/views/ResultView.vue'),
+    },
   ],
 });
